feat: learn more modal
